refactor: lean down App.tsx
move all handlers and api calls to context

this is in prep for concurrent conversation support
