+++ src/core/context/types.ts
@@ -32,6 +32,9 @@ export type AppContextType = {
   handleModelChange: (modelName: SupportedModels) => void;
   executeOperation: ExecuteOperation;
   loadConversation: (convoHistory: ConversationHistory) => void;
+  handleReset: () => void;
+  handleChatCompletion: (value: string) => void;
+  handleCancel: () => void;
   walletStore: WalletStore;
   client: OpenAI | null;
   thinkingStore: TextStreamStore;
